redid config files to use 8 bit fields instead of 32 bit (needs some work on sound settings but we'll handle that when sound is added). also updated f_wipe to be fully 16bit, made some more changes to various 32 bit fields that were tied to defaults. this was pretty far-reaching but i dont see any immediate bugs

+++ doomstat.h
@@ -96,8 +96,8 @@ extern  boolean         respawnmonsters;
 extern int32_t snd_MusicDevice;
 extern int32_t snd_SfxDevice;
 // Config file? Same disclaimer as above.
-extern int32_t snd_DesiredMusicDevice;
-extern int32_t snd_DesiredSfxDevice;
+extern uint8_t snd_DesiredMusicDevice;
+extern uint8_t snd_DesiredSfxDevice;
 
 
 // -------------------------
@@ -224,7 +224,7 @@ extern  boolean         precache;
 //  to force a wipe on the next draw
 extern  gamestate_t     wipegamestate;
 
-extern  int32_t             mouseSensitivity;
+extern  uint8_t             mouseSensitivity;
 //?
 // debug flag to cancel adaptiveness
 extern  boolean         singletics;
