+++ v_video.h
@@ -47,7 +47,7 @@ extern   MEMREF     screen4Ref;
 extern  int16_t	dirtybox[4];
 
 extern	byte	gammatable[5][256];
-extern	int32_t	usegamma;
+extern	uint8_t	usegamma;
